assessment/requirements.txt
Django==4.2.23
SpeechRecognition==3.10.1
pydub==0.25.1
python-Levenshtein==0.25.1